Se realiza un ajuste en el backend de Medicamentos

+++ src/main/java/com/uniminuto/clinica/repository/MedicamentoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MedicamentoRepository extends JpaRepository<Medicamento, Integer> {
+public interface MedicamentoRepository extends JpaRepository<Medicamento, Long> {
     Optional<Medicamento> findByNombre(String nombre);
 }
